# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "backen server of nero app. https://github.com/Furqanalimir/nero",
        "title": "Nero application docs [user and orders api]",
        "contact": {
            "name": "API Support",
            "url": "https://furqanali.vercel.app/",
            "email": "mrifurqan89@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:5050",
    "paths": {
        "/order": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Save order data into databae",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orders"
                ],
                "summary": "CreateOrder",
                "parameters": [
                    {
                        "description": "create order",
                        "name": "order",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/forms.OrderSwaggerForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "create response",
                        "schema": {
                            "$ref": "#/definitions/forms.ReqResSwagger"
                        }
                    },
                    "400": {
                        "description": "error response",
                        "schema": {
                            "$ref": "#/definitions/forms.ReqResSwagger"
                        }
                    }
                }
            }
        },
        "/order/:id": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "fetch user order by order id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orders"
                ],
                "summary": "fet order from database",
                "parameters": [
                    {
                        "type": "string",
                        "description": "order id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success response",
                        "schema": {
                            "$ref": "#/definitions/models.Order"
                        }
                    },
                    "400": {
                        "description": "error response",
                        "schema": {
                            "$ref": "#/definitions/forms.ReqResSwagger"
                        }
                    }
                }
            }
        },
        "/user/login": {
            "post": {
                "description": "validate user and get token",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "login user",
                "parameters": [
                    {
                        "description": "login user",
                        "name": "credentials",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/forms.Authenticate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "login response",
                        "schema": {
                            "$ref": "#/definitions/forms.ReqResSwagger"
                        }
                    },
                    "400": {
                        "description": "error response",
                        "schema": {
                            "$ref": "#/definitions/forms.ReqResSwagger"
                        }
                    }
                }
            }
        },
        "/user/signup": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Add user data to database",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Add User",
                "parameters": [
                    {
                        "description": "user info",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "signup response",
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    },
                    "400": {
                        "description": "error response",
                        "schema": {
                            "$ref": "#/definitions/forms.ReqResSwagger"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "forms.Authenticate": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string"
                },
                "phone": {
                    "type": "integer"
                }
            }
        },
        "forms.InfoSwaggerForm": {
            "type": "object",
            "properties": {
                "price": {
                    "type": "integer"
                },
                "productId": {
                    "type": "string"
                },
                "productName": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                }
            }
        },
        "forms.OrderSwaggerForm": {
            "type": "object",
            "properties": {
                "currency": {
                    "type": "string"
                },
                "info": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/forms.InfoSwaggerForm"
                    }
                },
                "tax": {
                    "type": "number"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "forms.ReqResSwagger": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/forms.data"
                },
                "error": {
                    "type": "string"
                }
            }
        },
        "forms.data": {
            "type": "object"
        },
        "models.Info": {
            "type": "object",
            "required": [
                "price",
                "product_id",
                "product_name"
            ],
            "properties": {
                "price": {
                    "type": "integer"
                },
                "product_id": {
                    "type": "string"
                },
                "product_name": {
                    "type": "string",
                    "minLength": 3
                },
                "quantity": {
                    "type": "integer"
                }
            }
        },
        "models.Order": {
            "type": "object",
            "required": [
                "currency",
                "tax",
                "total",
                "user_id"
            ],
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "currency": {
                    "type": "string"
                },
                "info": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Info"
                    }
                },
                "order_id": {
                    "type": "string"
                },
                "tax": {
                    "type": "number"
                },
                "total": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "models.User": {
            "type": "object",
            "required": [
                "age",
                "email",
                "gender",
                "password",
                "phone"
            ],
            "properties": {
                "active": {
                    "type": "boolean"
                },
                "age": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "integer"
                },
                "dob": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "gender": {
                    "type": "string",
                    "enum": [
                        "male",
                        "female"
                    ]
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string",
                    "minLength": 3
                },
                "password": {
                    "type": "string",
                    "minLength": 8
                },
                "phone": {
                    "type": "integer",
                    "minimum": 10
                },
                "profile_url": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}